implemented simulation result

+++ app/src/main/java/com/feliperce/investmentsimulator/data/remote/response/InvestmentParameter.kt
@@ -1,11 +1,13 @@
 package com.feliperce.investmentsimulator.data.remote.response
 
+import java.math.BigDecimal
+
 data class InvestmentParameter(
-	val rate: Double? = null,
+	val rate: BigDecimal? = null,
 	val maturityDate: String? = null,
 	val maturityBusinessDays: Int? = null,
-	val investedAmount: Double? = null,
+	val investedAmount: BigDecimal? = null,
 	val maturityTotalDays: Int? = null,
 	val isTaxFree: Boolean? = null,
-	val yearlyInterestRate: Double? = null
+	val yearlyInterestRate: BigDecimal? = null
 )
